+++ src/boot.js
@@ -1,9 +1,9 @@
-const Config = require('./config.js')
+const Config = require('./config/config.js')
 const GameServer = require('./server/server.js');
 
 // add to global namespace to reuse within deeper modules
 global.config = new Config();
-global.config.events.on('loaded', ()=> {
+global.config.ready(()=> {
   const server = new GameServer();
-  server.start(global.config.server.port);
-})
+  server.start(global.config.game.server.port);
+});
