+++ src/ui/script/content/me.js
@@ -1,18 +1,19 @@
 var Maumau = (typeof Maumau === "undefined" || !Maumau ) ? {} : Maumau;
 Maumau.Me = function(){
-  this.getName = function(ajax, callback){
+  this.getName = function(callback){
     // TODO: save custom name in local storage
     // if none - take local IP
-    ajax.get({
+    Maumau.xhr.get({
       path: '/api/ip',
       onSuccess: callback
     });
   }
 }
-// static function
+Maumau.me = new Maumau.Me();
+
+// static function for dynamic clientScript handler
 Maumau.Me.getName = function(callback){
-  const me = new Maumau.Me();
-  me.getName(new Maumau.Ajax(), name => {
+  Maumau.me.getName(name => {
     callback(`<span>${name}</span>`);
   });
 }
